refactor: console 관련 패키지 구조 일부 수정
